Fix documentation of HISTIGNORE
Patterns in HISTIGNORE result in commands matching these patterns _not_ to be saved on the history list. (as one would expect from the name of the variable)

Therefore, the current text `list of patterns used to decide which command lines should be saved on the history list.`, while not strictly/logically wrong, is vague and misleading.